ok there is actually a logic flaw in the running of trials in multiple threads with code coverage, so you will want to run the trials in a single thread when your asserting code coverage.
I am trying to think of a way around this, 
it would need to require that trials of 
dependent classes have executed before other trials,
but now sure how to queue this exactly.

The setup stuff is designed to run in multiple threads though,
this is a surprise to me.